+++ src/program/ui/InputEditorView.cpp
@@ -294,28 +294,28 @@ void InputEditorView::mainMenu(QPoint pos)
 
 void InputEditorView::insertInput()
 {
-    const QModelIndex index = selectionModel()->currentIndex();
-
+    const QModelIndexList indexes = selectionModel()->selectedRows();
+    
     /* If no row was selected, return */
-    if (!index.isValid())
+    if (indexes.count() == 0)
         return;
 
-    inputEditorModel->insertRows(index.row(), 1);
+    inputEditorModel->insertRows(indexes[0].row(), 1);
 }
 
 void InputEditorView::insertInputs()
 {
-    const QModelIndex index = selectionModel()->currentIndex();
+    const QModelIndexList indexes = selectionModel()->selectedRows();
 
     /* If no row was selected, return */
-    if (!index.isValid())
+    if (indexes.count() == 0)
         return;
 
     bool ok;
     int nbFrames = QInputDialog::getInt(this, tr("Insert frames"), tr("Number of frames to insert: "), 1, 0, 100000, 1, &ok);
 
     if (ok) {
-        inputEditorModel->insertRows(index.row(), nbFrames);
+        inputEditorModel->insertRows(indexes[0].row(), nbFrames);
     }
 }
 
@@ -343,15 +343,15 @@ void InputEditorView::deleteInput()
 
 void InputEditorView::truncateInputs()
 {
-    const QModelIndex index = selectionModel()->currentIndex();
+    const QModelIndexList indexes = selectionModel()->selectedRows();
 
     /* If no row was selected, return */
-    if (!index.isValid())
+    if (indexes.count() == 0)
         return;
 
     int nbRows = inputEditorModel->rowCount();
 
-    inputEditorModel->removeRows(index.row()+1, nbRows-index.row()-1);
+    inputEditorModel->removeRows(indexes[0].row()+1, nbRows-indexes[0].row()-1);
 }
 
 void InputEditorView::clearInput()
@@ -399,34 +399,34 @@ void InputEditorView::cutInputs()
 
 void InputEditorView::pasteInputs()
 {
-    const QModelIndex index = selectionModel()->currentIndex();
+    const QModelIndexList indexes = selectionModel()->selectedRows();
 
     /* If no row was selected, return */
-    if (!index.isValid())
+    if (indexes.count() == 0)
         return;
 
-    int nbFrames = inputEditorModel->pasteInputs(index.row());
+    int nbFrames = inputEditorModel->pasteInputs(indexes[0].row());
 
     /* Select the pasted inputs */
-    QModelIndex top = inputEditorModel->index(index.row(), 0);
-    QModelIndex bottom = inputEditorModel->index(index.row()+nbFrames-1, 0);
+    QModelIndex top = inputEditorModel->index(indexes[0].row(), 0);
+    QModelIndex bottom = inputEditorModel->index(indexes[0].row()+nbFrames-1, 0);
     selectionModel()->clear();
     selectionModel()->select(QItemSelection(top, bottom), QItemSelectionModel::Select | QItemSelectionModel::Rows);
 }
 
 void InputEditorView::pasteInsertInputs()
 {
-    const QModelIndex index = selectionModel()->currentIndex();
+    const QModelIndexList indexes = selectionModel()->selectedRows();
 
     /* If no row was selected, return */
-    if (!index.isValid())
+    if (indexes.count() == 0)
         return;
 
-    int nbFrames = inputEditorModel->pasteInsertInputs(index.row());
+    int nbFrames = inputEditorModel->pasteInsertInputs(indexes[0].row());
 
     /* Select the pasted inputs */
-    QModelIndex top = inputEditorModel->index(index.row(), 0);
-    QModelIndex bottom = inputEditorModel->index(index.row()+nbFrames-1, 0);
+    QModelIndex top = inputEditorModel->index(indexes[0].row(), 0);
+    QModelIndex bottom = inputEditorModel->index(indexes[0].row()+nbFrames-1, 0);
     selectionModel()->clear();
     selectionModel()->select(QItemSelection(top, bottom), QItemSelectionModel::Select | QItemSelectionModel::Rows);
 }